+++ src/instance/instance.js
@@ -8,6 +8,7 @@ import defaults from './defaults';
 // Import instance methods.
 import addLayer, { getLayerByName } from './methods/layer';
 import addPopup from './methods/popup';
+import { transform, transformExtent } from './methods/projection';
 import { zoomToVectors, zoomToLayer } from './methods/zoom';
 import { addBehavior, attachBehavior, attachBehaviorsByWeight } from './methods/behavior';
 import { measureGeometry } from '../utils/measure';
@@ -57,6 +58,8 @@ const createInstance = ({ target, options = {} }) => {
     attachBehavior,
     attachBehaviorsByWeight,
     measureGeometry,
+    transform,
+    transformExtent,
   };
 
   return instance;