+++ src/instance/methods/projection.js
@@ -1,0 +1,2 @@
+// Export OpenLayers transform functions
+export { transform, transformExtent } from 'ol/proj';