+++ CHANGELOG.md
@@ -6,6 +6,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Include transform and transformExtent OL methods. #171
+
 ## [v2.1.0] - 2022-11-15
 
 ### Added